refact: 파서 분리
- 파서 클래스 생성
